Revert "Refactorization + to_template()"
