--- cornflow_client/tests/data/gc_input.json
@@ -1,1 +1,0 @@
-{"pairs": [{"n1": 0, "n2": 5}, {"n1": 0, "n2": 15}, {"n1": 0, "n2": 32}, {"n1": 0, "n2": 44}, {"n1": 0, "n2": 47}, {"n1": 1, "n2": 4}, {"n1": 1, "n2": 22}, {"n1": 1, "n2": 47}, {"n1": 2, "n2": 5}, {"n1": 2, "n2": 7}, {"n1": 2, "n2": 14}, {"n1": 2, "n2": 22}, {"n1": 2, "n2": 30}, {"n1": 2, "n2": 33}, {"n1": 2, "n2": 42}, {"n1": 3, "n2": 7}, {"n1": 3, "n2": 11}, {"n1": 3, "n2": 12}, {"n1": 3, "n2": 25}, {"n1": 3, "n2": 47}, {"n1": 3, "n2": 49}, {"n1": 4, "n2": 13}, {"n1": 4, "n2": 24}, {"n1": 4, "n2": 45}, {"n1": 5, "n2": 14}, {"n1": 5, "n2": 41}, {"n1": 5, "n2": 44}, {"n1": 6, "n2": 8}, {"n1": 6, "n2": 33}, {"n1": 6, "n2": 49}, {"n1": 7, "n2": 8}, {"n1": 7, "n2": 18}, {"n1": 7, "n2": 30}, {"n1": 7, "n2": 41}, {"n1": 7, "n2": 43}, {"n1": 7, "n2": 47}, {"n1": 8, "n2": 10}, {"n1": 8, "n2": 26}, {"n1": 8, "n2": 32}, {"n1": 8, "n2": 36}, {"n1": 8, "n2": 39}, {"n1": 8, "n2": 47}, {"n1": 9, "n2": 24}, {"n1": 9, "n2": 48}, {"n1": 10, "n2": 18}, {"n1": 10, "n2": 24}, {"n1": 10, "n2": 32}, {"n1": 10, "n2": 40}, {"n1": 10, "n2": 41}, {"n1": 10, "n2": 47}, {"n1": 11, "n2": 13}, {"n1": 11, "n2": 19}, {"n1": 11, "n2": 27}, {"n1": 11, "n2": 36}, {"n1": 11, "n2": 48}, {"n1": 11, "n2": 49}, {"n1": 12, "n2": 23}, {"n1": 12, "n2": 24}, {"n1": 12, "n2": 41}, {"n1": 13, "n2": 33}, {"n1": 14, "n2": 23}, {"n1": 14, "n2": 26}, {"n1": 14, "n2": 45}, {"n1": 15, "n2": 21}, {"n1": 15, "n2": 25}, {"n1": 15, "n2": 29}, {"n1": 15, "n2": 48}, {"n1": 16, "n2": 18}, {"n1": 16, "n2": 19}, {"n1": 16, "n2": 24}, {"n1": 16, "n2": 27}, {"n1": 16, "n2": 30}, {"n1": 16, "n2": 39}, {"n1": 16, "n2": 46}, {"n1": 17, "n2": 19}, {"n1": 17, "n2": 20}, {"n1": 17, "n2": 23}, {"n1": 17, "n2": 44}, {"n1": 17, "n2": 49}, {"n1": 18, "n2": 22}, {"n1": 18, "n2": 30}, {"n1": 18, "n2": 37}, {"n1": 19, "n2": 26}, {"n1": 19, "n2": 36}, {"n1": 19, "n2": 38}, {"n1": 19, "n2": 41}, {"n1": 19, "n2": 46}, {"n1": 20, "n2": 30}, {"n1": 20, "n2": 40}, {"n1": 21, "n2": 35}, {"n1": 21, "n2": 40}, {"n1": 21, "n2": 48}, {"n1": 22, "n2": 34}, {"n1": 22, "n2": 35}, {"n1": 22, "n2": 37}, {"n1": 23, "n2": 39}, {"n1": 25, "n2": 37}, {"n1": 25, "n2": 38}, {"n1": 25, "n2": 46}, {"n1": 25, "n2": 49}, {"n1": 26, "n2": 27}, {"n1": 26, "n2": 43}, {"n1": 26, "n2": 47}, {"n1": 27, "n2": 42}, {"n1": 27, "n2": 43}, {"n1": 27, "n2": 46}, {"n1": 27, "n2": 47}, {"n1": 27, "n2": 49}, {"n1": 28, "n2": 43}, {"n1": 28, "n2": 44}, {"n1": 29, "n2": 40}, {"n1": 29, "n2": 45}, {"n1": 30, "n2": 38}, {"n1": 30, "n2": 40}, {"n1": 30, "n2": 47}, {"n1": 31, "n2": 32}, {"n1": 31, "n2": 35}, {"n1": 32, "n2": 36}, {"n1": 32, "n2": 44}, {"n1": 32, "n2": 48}, {"n1": 33, "n2": 34}, {"n1": 33, "n2": 40}, {"n1": 35, "n2": 39}, {"n1": 37, "n2": 48}, {"n1": 38, "n2": 48}, {"n1": 39, "n2": 46}, {"n1": 40, "n2": 43}, {"n1": 41, "n2": 49}, {"n1": 44, "n2": 45}, {"n1": 44, "n2": 47}, {"n1": 45, "n2": 49}, {"n1": 46, "n2": 47}, {"n1": 47, "n2": 49}]}
