--- cornflow_client/tests/data/gc_output.json
@@ -1,1 +1,0 @@
-{"assignment": [{"node": 0, "color": 1}, {"node": 1, "color": 0}, {"node": 2, "color": 1}, {"node": 3, "color": 1}, {"node": 4, "color": 1}, {"node": 5, "color": 0}, {"node": 6, "color": 0}, {"node": 7, "color": 0}, {"node": 8, "color": 1}, {"node": 9, "color": 1}, {"node": 10, "color": 2}, {"node": 11, "color": 3}, {"node": 12, "color": 2}, {"node": 13, "color": 0}, {"node": 14, "color": 3}, {"node": 15, "color": 0}, {"node": 16, "color": 3}, {"node": 17, "color": 1}, {"node": 18, "color": 1}, {"node": 19, "color": 0}, {"node": 20, "color": 0}, {"node": 21, "color": 1}, {"node": 22, "color": 2}, {"node": 23, "color": 0}, {"node": 24, "color": 0}, {"node": 25, "color": 3}, {"node": 26, "color": 2}, {"node": 27, "color": 0}, {"node": 28, "color": 0}, {"node": 29, "color": 1}, {"node": 30, "color": 2}, {"node": 31, "color": 1}, {"node": 32, "color": 0}, {"node": 33, "color": 2}, {"node": 34, "color": 0}, {"node": 35, "color": 3}, {"node": 36, "color": 2}, {"node": 37, "color": 0}, {"node": 38, "color": 1}, {"node": 39, "color": 2}, {"node": 40, "color": 3}, {"node": 41, "color": 1}, {"node": 42, "color": 2}, {"node": 43, "color": 1}, {"node": 44, "color": 2}, {"node": 45, "color": 0}, {"node": 46, "color": 1}, {"node": 47, "color": 3}, {"node": 48, "color": 2}, {"node": 49, "color": 2}]}
