+++ cornflow_client/tests/data/graph_coloring2_input.json
@@ -11,16 +11,10 @@
         "pairs": {
             "type": "array",
             "items": {
-                "type": "object",
-                "properties": {
-                    "n1": {
-                        "type": "integer"
-                    },
-                    "n2": {
-                        "type": "integer"
-                    }
-                },
-                "required": ["n1", "n2"]
+                "type": "array",
+                "items": {
+                    "type": "integer"
+                }
             }
         }
     },
