--- cornflow_client/tests/data/graph_coloring_output.json
@@ -1,27 +1,0 @@
-{
-	"$schema": "http://json-schema.org/schema#",
-	"type": "object",
-	"properties": {
-		"assignment": {
-			"type": "array",
-			"items": {
-				"type": "object",
-				"properties": {
-					"node": {
-						"type": "integer"
-					},
-					"color": {
-						"type": "integer"
-					}
-				},
-				"required": [
-					"color",
-					"node"
-				]
-			}
-		}
-	},
-	"required": [
-		"assignment"
-	]
-}
