+++ cornflow_client/tests/data/hk_data_schema.json
@@ -3,7 +3,6 @@
 	"type": "object",
 	"properties": {
 		"durations": {
-			"description": "The duration of a job when in some mode",
 			"type": "array",
 			"items": {
 				"type": "object",
@@ -18,11 +17,10 @@
 						"type": "integer"
 					}
 				},
-				"required": ["job", "mode", "duration"]
+				"required": ["duration", "job", "mode"]
 			}
 		},
 		"jobs": {
-			"description": "Successor job ids for each job id",
 			"type": "array",
 			"items": {
 				"type": "object",
@@ -41,7 +39,6 @@
 			}
 		},
 		"needs": {
-			"description": "Resource consumption (need) for each resource when doing a job in some mode",
 			"type": "array",
 			"items": {
 				"type": "object",
@@ -59,11 +56,10 @@
 						"type": "string"
 					}
 				},
-				"required": ["job", "mode", "resource", "need"]
+				"required": ["job", "mode", "need", "resource"]
 			}
 		},
 		"resources": {
-			"description": "Total quantity of available resource for each resource id",
 			"type": "array",
 			"items": {
 				"type": "object",
@@ -75,7 +71,7 @@
 						"type": "string"
 					}
 				},
-				"required": ["id", "available"]
+				"required": ["available", "id"]
 			}
 		}
 	},
