+++ requirements-dev.txt
@@ -1,3 +1,2 @@
 -r requirements.txt
 pulp
-pandas
